Server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
openai==0.28.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.3.0
psutil==6.0.0
py-cpuinfo==9.0.0
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sympy==1.12.1
tbb==2021.13.0
thop>=0.1.1.post2209072238
torch==1.12.0
torchvision==0.13.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.42
ultralytics-thop==2.0.0
urllib3==2.2.2
utils==1.0.2
yarl==1.9.4
zipp==3.19.2
